Safari-611.1.10/LayoutTests/plugins/netscape-destroy-plugin-script-objects-expected.txt

CONSOLE MESSAGE: PLUGIN: NPP_Destroy
This tests that accessing plug-in script objects from a destroyed plugin cause throws the right exceptions.
SUCCESS
